The test application and the plgind error messages has been improved.
Original commit message from CVS:
The test application and the plgind error messages has been improved.

+++ sys/v4l2/gstv4l2src.h
@@ -99,7 +99,7 @@ struct _GstV4l2Src
   /* True if we want to stop */
   gboolean quit, is_capturing;
 
-  gint offset;
+  guint64 offset;
 
   /* how are we going to push buffers? */
   guint fps_n, fps_d;
